+++ lib/qtl-data.js
@@ -1,10 +1,12 @@
+import _ from "lodash";
+import { ASSOCIATED_ORGAN_MAP } from "./chromatin-data";
 import {
   TissueScoreHexColor,
   TissueScoreTailwindColor,
 } from "./tissue-specific-score";
 
 /**
- * This function get filterred QTL data
+ * This function get filtered QTL data
  * @param {Array} data all the QTL data
  * @param {Array} organFilters a list of organs for filtering the data
  * @returns a list of QTL datasets
@@ -16,14 +18,7 @@ export function getFilteredQtlData(data, organFilters) {
       const organsForDataset = dataset.biosample_ontology?.organ_slims
         ? dataset.biosample_ontology.organ_slims
         : [];
-      let hasOrgan = false;
-      for (let j = 0; j < organsForDataset.length; j++) {
-        if (organFilters.includes(organsForDataset[j])) {
-          hasOrgan = true;
-          break;
-        }
-      }
-      return hasOrgan;
+      return organsForDataset.some((organ) => organFilters.includes(organ));
     });
   }
   return filteredData;
@@ -71,4 +66,49 @@ export function getFillColorTailwind(facets, organ) {
 export function getFillColorHex(facets, organ) {
   const fill = organ in facets ? TissueScoreHexColor[facets[organ]] : "white";
   return fill;
+}
+
+/**
+ * @param {*} organFilters before change
+ * @param {*} organ the user clicks
+ * @param {*} organList shown in body map
+ * @param {*} enabledOrganList enabled for selecting
+ * @returns the updated organ filter
+ */
+export function getOrganFilter(
+  organFilters,
+  organ,
+  organList,
+  enabledOrganList
+) {
+  let filters = [...organFilters];
+  if (organList && enabledOrganList) {
+    if (organList.includes(organ) && enabledOrganList.includes(organ)) {
+      if (filters.includes(organ)) {
+        if (organ in ASSOCIATED_ORGAN_MAP) {
+          ASSOCIATED_ORGAN_MAP[organ].forEach((associatedOrgan) => {
+            if (enabledOrganList.includes(associatedOrgan)) {
+              filters = _.without(filters, associatedOrgan);
+            }
+          });
+        }
+        filters = _.without(filters, organ);
+      } else {
+        filters.push(organ);
+        if (organ in ASSOCIATED_ORGAN_MAP) {
+          ASSOCIATED_ORGAN_MAP[organ].forEach((associatedOrgan) => {
+            if (
+              enabledOrganList.includes(associatedOrgan) &&
+              !filters.includes(associatedOrgan)
+            ) {
+              filters.push(associatedOrgan);
+            }
+          });
+        }
+      }
+    }
+  } else {
+    filters = _.without(filters, organ);
+  }
+  return filters;
 }
