+++ lib/tissue-specific-score.js
@@ -55,7 +55,7 @@ export function getNormalizedTissueSpecificScore(tissueSpecificScores) {
 /**
  * @param {object} data contains the datasets to update
  * @param {object} normalizedTissueSpecificScore to look up normalized tissue specific score
- * @returns data with tisssue specific score in each dataset
+ * @returns data with tissue specific score in each dataset
  */
 export function getDataWithTissueScore(data, normalizedTissueSpecificScore) {
   const datasets = data["@graph"];
